+++ tests/test_imports/test_import.py
@@ -2,6 +2,7 @@ import logging
 import unittest
 from satella.imports import import_class
 import subprocess
+import warnings
 
 logger = logging.getLogger(__name__)
 
@@ -9,7 +10,9 @@ logger = logging.getLogger(__name__)
 class TestImports(unittest.TestCase):
     def test_imports(self):
         import tests.test_imports.importa
-        tests.test_imports.importa.do_import()
+        with warnings.catch_warnings() as warns:
+            tests.test_imports.importa.do_import()
+        self.assertGreater(len(warns), 0)
 
         # this as well checks for the namespace's pollution
         self.assertEqual(set(tests.test_imports.importa.importb.__all__),